requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-knox==5.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pika==1.3.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
